tried a first version of "predictor validity".  It's not strong enough. in addition to saying that (predictor fuel) is finite for each fuel, we also need to say that there's a large enough F s.t. (predictor F) = (predictor fuel) for all fuel >= F.
To prove this (about the F), we need to bound the fuel required for the
spilled trace in terms of only the source trace and the source program -
that is, without referencing any proof about the source program.  This is
necessary, since any proof about the source program would depend on
the starting memory and trace of the program, and we need F to be
independent of these.

I'm not totally convinced by that argument, so I'll try again.  Look at
the lemma 'last_step' in Pipeline.v.  That's the sort of thing we need -
that the amount of fuel required for the predictor function is dependent
only on the predictor function.  It can't be dependent on a hypothesis of
the form 'FlatToRiscvCommon.runsto initial post', as in runsTo_sane, since
we need some assumptions about the starting trace and memory to get to this
hypothesis.  So, as mentioned above, we really do need to bound the fuel
required only in terms of the source trace and the source program.
This makes a lot of sense - we've gone to so much effort to show that the
target-level predictor function depends only on the source trace and
source program, it only makes sense that we'd need to prove the same thing
about the amount of fuel we need to give to the target predictor function.

Of course, this whole issue goes away if we use excluded middle.